Migration of sheet generation to Sheet classes
As part of migrating assay-type checks from SequencingProcess and to
PoolComposition, sample sheet and prep info sheet generation has been
broken up into several classes to help facilitate that process more
easily.
